add layout context for diff layouts based on authorized state
