+++ client/src/app/home/page.tsx
@@ -1,173 +1,5 @@
 "use client";
 
-import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import DashboardIcon from "@mui/icons-material/Dashboard";
-import LogoutIcon from "@mui/icons-material/Logout";
-import MenuIcon from "@mui/icons-material/Menu";
-import {
-	Box,
-	Container,
-	Divider,
-	IconButton,
-	List,
-	ListItemButton,
-	ListItemIcon,
-	ListItemText,
-	Toolbar,
-	Typography,
-	styled,
-} from "@mui/material";
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
-import MuiDrawer from "@mui/material/Drawer";
-import axios from "axios";
-import { usePathname, useRouter } from "next/navigation";
-import React from "react";
-
-interface AppBarProps extends MuiAppBarProps {
-	open?: boolean;
-}
-
-const drawerWidth: number = 240;
-
-const AppBar = styled(MuiAppBar, {
-	shouldForwardProp: (prop) => prop !== "open",
-})<AppBarProps>(({ theme, open }) => ({
-	zIndex: theme.zIndex.drawer + 1,
-	transition: theme.transitions.create(["width", "margin"], {
-		easing: theme.transitions.easing.sharp,
-		duration: theme.transitions.duration.leavingScreen,
-	}),
-	...(open && {
-		marginLeft: drawerWidth,
-		width: `calc(100% - ${drawerWidth}px)`,
-		transition: theme.transitions.create(["width", "margin"], {
-			easing: theme.transitions.easing.sharp,
-			duration: theme.transitions.duration.enteringScreen,
-		}),
-	}),
-}));
-
-const Drawer = styled(MuiDrawer, {
-	shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }) => ({
-	"& .MuiDrawer-paper": {
-		position: "relative",
-		whiteSpace: "nowrap",
-		width: drawerWidth,
-		transition: theme.transitions.create("width", {
-			easing: theme.transitions.easing.sharp,
-			duration: theme.transitions.duration.enteringScreen,
-		}),
-		boxSizing: "border-box",
-		...(!open && {
-			overflowX: "hidden",
-			transition: theme.transitions.create("width", {
-				easing: theme.transitions.easing.sharp,
-				duration: theme.transitions.duration.leavingScreen,
-			}),
-			width: theme.spacing(7),
-			[theme.breakpoints.up("sm")]: {
-				width: theme.spacing(9),
-			},
-		}),
-	},
-}));
-
 export default function Home() {
-	const router = useRouter();
-	const pathname = usePathname();
-	const [open, setOpen] = React.useState(false);
-	const toggleDrawer = () => {
-		setOpen(!open);
-	};
-
-	const onClickLogout = async () => {
-		const res = await axios
-			.post("http://localhost:8080/auth/logout", {}, { withCredentials: true })
-			.then((res) => res.data);
-		if (res.event === "USER_LOGOUT_SUCCESS") {
-			router.push("/");
-		}
-	};
-
-	return (
-		<Box sx={{ display: "flex" }}>
-			<AppBar position="absolute" open={open}>
-				<Toolbar
-					sx={{
-						pr: "24px", // keep right padding when drawer closed
-					}}
-				>
-					<IconButton
-						edge="start"
-						color="inherit"
-						aria-label="open drawer"
-						onClick={toggleDrawer}
-						sx={{
-							marginRight: "36px",
-							...(open && { display: "none" }),
-						}}
-					>
-						<MenuIcon />
-					</IconButton>
-					<Typography
-						component="h1"
-						variant="h6"
-						color="inherit"
-						noWrap
-						sx={{ flexGrow: 1 }}
-					>
-						Jobdash
-					</Typography>
-				</Toolbar>
-			</AppBar>
-			<Drawer variant="permanent" open={open}>
-				<Toolbar
-					sx={{
-						display: "flex",
-						alignItems: "center",
-						justifyContent: "flex-end",
-						px: [1],
-					}}
-				>
-					<IconButton onClick={toggleDrawer}>
-						<ChevronLeftIcon />
-					</IconButton>
-				</Toolbar>
-				<Divider />
-				<List component="nav">
-					<ListItemButton selected={pathname === "/home"}>
-						<ListItemIcon>
-							<DashboardIcon />
-						</ListItemIcon>
-						<ListItemText primary="Dashboard" />
-					</ListItemButton>
-					<Divider sx={{ my: 1 }} />
-					<ListItemButton onClick={onClickLogout}>
-						<ListItemIcon>
-							<LogoutIcon />
-						</ListItemIcon>
-						<ListItemText primary="Logout" />
-					</ListItemButton>
-				</List>
-			</Drawer>
-			<Box
-				component="main"
-				sx={{
-					backgroundColor: (theme) =>
-						theme.palette.mode === "light"
-							? theme.palette.grey[100]
-							: theme.palette.grey[900],
-					flexGrow: 1,
-					height: "100vh",
-					overflow: "auto",
-				}}
-			>
-				<Toolbar />
-				<Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-					Hello world this is where my app will be
-				</Container>
-			</Box>
-		</Box>
-	);
+	return <div>Hello world</div>;
 }
